cuong

+++ WebSite/view/user/list_user.php
@@ -73,13 +73,14 @@
 					<td><?php echo $value['tokenExpire'] ?></td>
 					
 					<td>
+						<a href="admin.php?controller=user&action=update&id=<?php echo $value['user_id']?>">Edit</a>
 						
-                           <a href="admin.php?controller=user&action=delete&id=<?php echo $value['user_id'] ?>"> Delete</a>
+                           
 					</td>
 					<td>	 
 
+                    <a onclick="return confirm('Are you want to delete record?')" href="admin.php?controller=user&action=delete&id=<?php echo $value['user_id'] ?>"> Delete</a>
                     
-                    <a href="admin.php?controller=user&action=update&id=<?php echo $value['user_id']?>">Edit</a>
 						<!-- <input type="submit" > -->
 					</td>
 				</tr>
